Refactored OpenTelemetry tests
