+++ integration_test_suite/integration_tests/src/test/java/io/prometheus/jmx/test/opentelemetry/AbstractOpenTelemetryTest.java
@@ -3,43 +3,26 @@ package io.prometheus.jmx.test.opentelemetry;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.fail;
 
-import com.github.dockerjava.api.model.Ulimit;
 import io.prometheus.jmx.test.support.JmxExporterMode;
-import io.prometheus.jmx.test.support.http.HttpClient;
 import io.prometheus.jmx.test.support.http.HttpRequest;
 import io.prometheus.jmx.test.support.http.HttpResponse;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
-import java.time.Duration;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
 import org.antublue.test.engine.api.TestEngine;
 import org.antublue.test.engine.extras.throttle.ExponentialBackoffThrottle;
 import org.antublue.test.engine.extras.throttle.Throttle;
-import org.testcontainers.containers.BindMode;
-import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.containers.Network;
-import org.testcontainers.containers.startupcheck.IsRunningStartupCheckStrategy;
-import org.testcontainers.containers.wait.strategy.Wait;
 import org.testcontainers.shaded.org.yaml.snakeyaml.Yaml;
 
 /** Class to implement AbstractOpenTelemetryTest */
 public abstract class AbstractOpenTelemetryTest {
 
-    private static final long MEMORY_BYTES = 1073741824; // 1 GB
-    private static final long MEMORY_SWAP_BYTES = 2 * MEMORY_BYTES;
-
-    private static final String BASE_URL = "http://localhost";
-
     private Network network;
-    private GenericContainer<?> prometheusContainer;
-    private GenericContainer<?> standaloneApplicationContainer;
-    private GenericContainer<?> javaAgentApplicationContainer;
-    private GenericContainer<?> standaloneExporterContainer;
-    private HttpClient httpClient;
 
-    @TestEngine.Argument public OpenTelemetryTestArguments openTelemetryTestArguments;
+    @TestEngine.Argument public OpenTelemetryTestEnvironment openTelemetryTestEnvironment;
 
     @TestEngine.Prepare
     public void prepare() {
@@ -50,57 +33,7 @@ public abstract class AbstractOpenTelemetryTest {
 
     @TestEngine.BeforeAll
     public void beforeAll() {
-        switch (openTelemetryTestArguments.getJmxExporterMode()) {
-            case JavaAgent:
-                {
-                    prometheusContainer =
-                            createPrometheusContainer(
-                                    openTelemetryTestArguments.getJmxExporterMode(),
-                                    network,
-                                    openTelemetryTestArguments.getPrometheusDockerImageName(),
-                                    getClass().getName());
-
-                    javaAgentApplicationContainer =
-                            createJavaAgentApplicationContainer(
-                                    network,
-                                    openTelemetryTestArguments.getJavaDockerImageName(),
-                                    getClass().getName());
-
-                    prometheusContainer.start();
-                    javaAgentApplicationContainer.start();
-
-                    break;
-                }
-            case Standalone:
-                {
-                    prometheusContainer =
-                            createPrometheusContainer(
-                                    openTelemetryTestArguments.getJmxExporterMode(),
-                                    network,
-                                    openTelemetryTestArguments.getPrometheusDockerImageName(),
-                                    getClass().getName());
-
-                    standaloneApplicationContainer =
-                            createStandaloneApplicationContainer(
-                                    network,
-                                    openTelemetryTestArguments.getJavaDockerImageName(),
-                                    getClass().getName());
-
-                    standaloneExporterContainer =
-                            createStandaloneExporterContainer(
-                                    network,
-                                    openTelemetryTestArguments.getJavaDockerImageName(),
-                                    getClass().getName());
-
-                    prometheusContainer.start();
-                    standaloneApplicationContainer.start();
-                    standaloneExporterContainer.start();
-
-                    break;
-                }
-        }
-
-        httpClient = createPrometheusHttpClient(prometheusContainer, BASE_URL, 9090);
+        openTelemetryTestEnvironment.initialize(getClass(), network);
     }
 
     /** Method to test that Prometheus is up */
@@ -150,7 +83,7 @@ public abstract class AbstractOpenTelemetryTest {
         ExpectedMetricsNames.getMetricsNames().stream()
                 .filter(
                         metricName -> {
-                            if (openTelemetryTestArguments.getJmxExporterMode()
+                            if (openTelemetryTestEnvironment.getJmxExporterMode()
                                                     == JmxExporterMode.Standalone
                                             && metricName.startsWith("jvm_")
                                     || metricName.startsWith("process_")) {
@@ -168,25 +101,7 @@ public abstract class AbstractOpenTelemetryTest {
 
     @TestEngine.AfterAll
     public void afterAll() {
-        if (javaAgentApplicationContainer != null) {
-            javaAgentApplicationContainer.close();
-            javaAgentApplicationContainer = null;
-        }
-
-        if (standaloneExporterContainer != null) {
-            standaloneExporterContainer.close();
-            standaloneExporterContainer = null;
-        }
-
-        if (standaloneApplicationContainer != null) {
-            standaloneApplicationContainer.close();
-            standaloneApplicationContainer = null;
-        }
-
-        if (prometheusContainer != null) {
-            prometheusContainer.close();
-            prometheusContainer = null;
-        }
+        openTelemetryTestEnvironment.destroy();
     }
 
     @TestEngine.Conclude
@@ -196,10 +111,23 @@ public abstract class AbstractOpenTelemetryTest {
         }
     }
 
+    /**
+     * Method to get a Prometheus metric
+     *
+     * @param metricName metricName
+     * @return the metric value, or null if it doesn't exist
+     */
     protected Double getPrometheusMetric(String metricName) {
         return getPrometheusMetric(metricName, null);
     }
 
+    /**
+     * Method to get a Promethehus metrics
+     *
+     * @param metricName metricName
+     * @param labels labels
+     * @return the metric value, or null if it doesn't exist
+     */
     protected Double getPrometheusMetric(String metricName, String[] labels) {
         Throttle throttle = new ExponentialBackoffThrottle(100, 6400);
         AtomicReference<Double> value = new AtomicReference<>();
@@ -247,198 +175,6 @@ public abstract class AbstractOpenTelemetryTest {
      * @return an HttpResponse
      */
     protected HttpResponse sendRequest(String path) {
-        return httpClient.send(new HttpRequest(path));
-    }
-
-    /**
-     * Method to create an HttpClient
-     *
-     * @param genericContainer genericContainer
-     * @param baseUrl baseUrl
-     * @return the return value
-     */
-    private static HttpClient createPrometheusHttpClient(
-            GenericContainer<?> genericContainer, String baseUrl, int mappedPort) {
-        return new HttpClient(baseUrl + ":" + genericContainer.getMappedPort(mappedPort));
-    }
-
-    /**
-     * Method to create a Prometheus container
-     *
-     * @param network network
-     * @param prometheusDockerImageName prometheusDockerImageName
-     * @param testName testName
-     * @return the return value
-     */
-    private static GenericContainer<?> createPrometheusContainer(
-            JmxExporterMode jmxExporterMode,
-            Network network,
-            String prometheusDockerImageName,
-            String testName) {
-        return new GenericContainer<>(prometheusDockerImageName)
-                .withClasspathResourceMapping(
-                        testName.replace(".", "/") + "/" + jmxExporterMode + "/prometheus.yml",
-                        "/etc/prometheus/prometheus.yml",
-                        BindMode.READ_ONLY)
-                .withWorkingDirectory("/prometheus")
-                .withCommand(
-                        "--config.file=/etc/prometheus/prometheus.yml",
-                        "--storage.tsdb.path=/prometheus",
-                        "--web.console.libraries=/usr/share/prometheus/console_libraries",
-                        "--web.console.templates=/usr/share/prometheus/consoles",
-                        "--enable-feature=otlp-write-receiver")
-                .withCreateContainerCmdModifier(
-                        c ->
-                                c.getHostConfig()
-                                        .withMemory(MEMORY_BYTES)
-                                        .withMemorySwap(MEMORY_SWAP_BYTES))
-                .withCreateContainerCmdModifier(
-                        c ->
-                                c.getHostConfig()
-                                        .withUlimits(
-                                                new Ulimit[] {
-                                                    new Ulimit("nofile", 65536L, 65536L)
-                                                }))
-                .withExposedPorts(9090)
-                .withLogConsumer(
-                        outputFrame -> {
-                            String string = outputFrame.getUtf8StringWithoutLineEnding().trim();
-                            if (!string.isBlank()) {
-                                System.out.println(string);
-                            }
-                        })
-                .withNetwork(network)
-                .withNetworkAliases("prometheus")
-                .withStartupCheckStrategy(new IsRunningStartupCheckStrategy())
-                .withStartupTimeout(Duration.ofMillis(30000));
-    }
-
-    /**
-     * Method to create an application container
-     *
-     * @param network network
-     * @param dockerImageName dockerImageName
-     * @param testName testName
-     * @return the return value
-     */
-    private static GenericContainer<?> createJavaAgentApplicationContainer(
-            Network network, String dockerImageName, String testName) {
-        return new GenericContainer<>(dockerImageName)
-                .waitingFor(Wait.forLogMessage(".*Running.*", 1))
-                .withClasspathResourceMapping("common", "/temp", BindMode.READ_ONLY)
-                .withClasspathResourceMapping(
-                        testName.replace(".", "/") + "/JavaAgent", "/temp", BindMode.READ_ONLY)
-                .withCreateContainerCmdModifier(
-                        c ->
-                                c.getHostConfig()
-                                        .withMemory(MEMORY_BYTES)
-                                        .withMemorySwap(MEMORY_SWAP_BYTES))
-                .withCreateContainerCmdModifier(
-                        c ->
-                                c.getHostConfig()
-                                        .withUlimits(
-                                                new Ulimit[] {
-                                                    new Ulimit("nofile", 65536L, 65536L)
-                                                }))
-                .withCommand("/bin/sh application.sh")
-                .withExposedPorts(8888)
-                .withLogConsumer(
-                        outputFrame -> {
-                            String string = outputFrame.getUtf8StringWithoutLineEnding().trim();
-                            if (!string.isBlank()) {
-                                System.out.println(string);
-                            }
-                        })
-                .withNetwork(network)
-                .withNetworkAliases("application")
-                .withStartupCheckStrategy(new IsRunningStartupCheckStrategy())
-                .withStartupTimeout(Duration.ofMillis(30000))
-                .withWorkingDirectory("/temp");
-    }
-
-    /**
-     * Method to create an application container
-     *
-     * @param network network
-     * @param dockerImageName dockerImageName
-     * @param testName testName
-     * @return the return value
-     */
-    private static GenericContainer<?> createStandaloneApplicationContainer(
-            Network network, String dockerImageName, String testName) {
-        return new GenericContainer<>(dockerImageName)
-                .waitingFor(Wait.forLogMessage(".*Running.*", 1))
-                .withClasspathResourceMapping("common", "/temp", BindMode.READ_ONLY)
-                .withClasspathResourceMapping(
-                        testName.replace(".", "/") + "/Standalone", "/temp", BindMode.READ_ONLY)
-                .withCreateContainerCmdModifier(
-                        c ->
-                                c.getHostConfig()
-                                        .withMemory(MEMORY_BYTES)
-                                        .withMemorySwap(MEMORY_SWAP_BYTES))
-                .withCreateContainerCmdModifier(
-                        c ->
-                                c.getHostConfig()
-                                        .withUlimits(
-                                                new Ulimit[] {
-                                                    new Ulimit("nofile", 65536L, 65536L)
-                                                }))
-                .withCommand("/bin/sh application.sh")
-                .withExposedPorts(9999)
-                .withLogConsumer(
-                        outputFrame -> {
-                            String string = outputFrame.getUtf8StringWithoutLineEnding().trim();
-                            if (!string.isBlank()) {
-                                System.out.println(string);
-                            }
-                        })
-                .withNetwork(network)
-                .withNetworkAliases("application")
-                .withStartupCheckStrategy(new IsRunningStartupCheckStrategy())
-                .withStartupTimeout(Duration.ofMillis(30000))
-                .withWorkingDirectory("/temp");
-    }
-
-    /**
-     * Method to create an exporter container
-     *
-     * @param network network
-     * @param dockerImageName dockerImageName
-     * @param testName testName
-     * @return the return value
-     */
-    private static GenericContainer<?> createStandaloneExporterContainer(
-            Network network, String dockerImageName, String testName) {
-        return new GenericContainer<>(dockerImageName)
-                .waitingFor(Wait.forListeningPort())
-                .withClasspathResourceMapping("common", "/temp", BindMode.READ_ONLY)
-                .withClasspathResourceMapping(
-                        testName.replace(".", "/") + "/Standalone", "/temp", BindMode.READ_ONLY)
-                .withCreateContainerCmdModifier(
-                        c ->
-                                c.getHostConfig()
-                                        .withMemory(MEMORY_BYTES)
-                                        .withMemorySwap(MEMORY_SWAP_BYTES))
-                .withCreateContainerCmdModifier(
-                        c ->
-                                c.getHostConfig()
-                                        .withUlimits(
-                                                new Ulimit[] {
-                                                    new Ulimit("nofile", 65536L, 65536L)
-                                                }))
-                .withCommand("/bin/sh exporter.sh")
-                .withExposedPorts(8888)
-                .withLogConsumer(
-                        outputFrame -> {
-                            String string = outputFrame.getUtf8StringWithoutLineEnding().trim();
-                            if (!string.isBlank()) {
-                                System.out.println(string);
-                            }
-                        })
-                .withNetwork(network)
-                .withNetworkAliases("exporter")
-                .withStartupCheckStrategy(new IsRunningStartupCheckStrategy())
-                .withStartupTimeout(Duration.ofMillis(30000))
-                .withWorkingDirectory("/temp");
+        return openTelemetryTestEnvironment.getPrometheusHttpClient().send(new HttpRequest(path));
     }
 }
